+++ internal/api/handlers_auth.go
@@ -9,9 +9,9 @@ import (
 
 const (
 	serverVersion                = "0.2.0"
-	apiVersion                   = "1.10.0"
+	apiVersion                   = "1.11.0"
 	minimumClientAPIVersion      = "1.0.0"
-	eventAPIVersion              = "1.11.0"
+	eventAPIVersion              = "1.12.0"
 	minimumClientEventAPIVersion = "1.3.0"
 )